This is C++!
- Fix header file names
- Fix include guard defines
